[FEATURE] allow multiple folders for static datastructures
Idea:
A templavoila datastructure is tightly coupled with it's template.
Although one could argue, that most templates might be "main content"
and "sidebar" content, this doesn't need to be the case and if merging
two systems, although the structure might be the same, the naming of the
fields might not be. Additionally a datasctructure might also support
really complex options and fields. Thus a datastructure and template
combination might be called unique. Of course for "one design" an
administrator will probably make a combined datastructure for all
templates.
Nevertheless a TYPO3 instance can hold multiple designs (e.g. per
domain). As it is best practice nowadays to define a web template as an
extension this means that the static datastructures should be supplied
in the very same extension.
This leads to the necessity to define more than one folder in extension
configuration, e.g. multiple paths separated by a semicolon (colon
perhaps breaks volume-path-separator on Windows), as using a
fileadmin/templates folder doesn't make sense then.

Solution:
As the path is used for selecting a DS and creating new DS, one path
needs to be writeable. The easiest way for this, would be to have the
first path "as is" and use additional paths just for DS selection. Added
this in the description of the config option

For all access to staticDS.path_page and .path_fce a
"getFirstDirectoryInList" is used.
For the pull-down menu for TOBJ all paths are used.

Resolves: #65