[BUGFIX] Fix introduced hook to own hook config
Using the core hook config means to also call the hook with correct
class, which isn't possible for us. So we rename the implemented hooks
from #79 to new names (as they were never released).
New namings/configs are:
$TYPO3_CONF_VARS['SC_OPTIONS']['templavoilaplus']['BackendLayout']['renderHeaderFunctionHook']
$TYPO3_CONF_VARS['SC_OPTIONS']['templavoilaplus']['BackendLayout']['renderFooterFunctionHook']

Beside this also cleaning the hooks should be done so following hooks
are replaced with deprecation.

$TYPO3_CONF_VARS['EXTCONF']['templavoilaplus']['mod1']['renderTopToolbar']
$TYPO3_CONF_VARS['EXTCONF']['templavoilaplus']['mod1']['render_editPageScreen']

to the new ones:

$TYPO3_CONF_VARS['SC_OPTIONS']['templavoilaplus']['BackendLayout']['renderEditPageHeaderFunctionHook']
$TYPO3_CONF_VARS['SC_OPTIONS']['templavoilaplus']['BackendLayout']['renderEditPageFooterFunctionHook']

Definition:
"A function Hook is a hook which calls the given function."
"A object Hook is a Hook which needs to implement a named method/attribute
which will be called/used, it will mostly be interfaced."

Related: #79
Resolves: #80