[BUGFIX] Make d'n'd of content elements a bit nicer
Drag'n'Drop into empty containers is sometimes very hard if not
impossible. This patch should help to get this done better. But the
issues are lying deeply in the jQueryUI sortable library. We can only
try to be a bit smarter to this JS. Which will be done with this commit.

Resolves: #87