[BUGFIX] Remove special char from extension name to help TER
In older TYPO3 versions the special char can't be read and so updating
the extension list will fail. For this a script was introduced which
removes the special chars inside TER. So changing the name from à to a
so the special char won't get removed and we lost our name.

Resolves: #94