[FEATURE] Add tt_content.layout as CSS class to BEmodule output
The Backend module outputs many HTML/CSS classes for content
elements according to the fields in the tt_content table (e.g.
CType, tx_templavoilaplus_tobj). Obviously the field 'layout'
should by definition have an optional impact on the output as well.

One could argue that this must not be the case and that each
arbitrary tt_content field might be relevant, however layout was
designed for this purpose in the first place. So I propose to add
this specific field to the output.

This would improve the option to style the backend module with a
custom CSS via the PageTSconfig
mod.web_txtemplavoilaplusLayout.stylesheet.myvalue  = path/to/my.css

Resolves: #182