[FEATURE] v9: Replace resolveSheetDefInDS/resolveAllSheetsInDS
We integrate this both functions in our own TemplaVoilaUtility class.
This isn't perfect but the resolving part inside FlexFormUtility isn't
generic and directly useable. So we would be in need todo more work on
this point. Which is contraproductive for other parts that need a
rewrite. So solving this with a TODO marker.

Resolves: #114