[REGRESSION] content_from_pid can't resolve map in all cases If the main page (uid 123) which defines the map has a subpage (uid 456) and uses "content_from_pid" to point to that subpage the page cannot be rendered.
The record is the subpage record (456), the rootline is the mainPage rootline. However while
getting map from rootline the first element is skipped as it is the current page in most cases.

This patch removes that skipping, as it is barely slower and allows to identify the parent map
in a content_from_pid case.