[TASK] make PageLayout compatible with <=10LTS
as 10LTS and below use bootstrap v4 instead of v5 the current page backend module design is broken.
Backporting the 11LTS backend.css and applying some overrides might not be the most elegant solution, however:
a) it works
b) it reduces future errors as only one version of the layout needs to be maintained.
Thus, I propose this PR to make it look pixel perfect on 10LTS.