[BUGFIX] ddev breaks processValue wraps
Using ddev 1.21.1 with php 7.4 my working config shows up broken. Multiple
wraps are missing or the label `__TEMPLAVOILAPLUS__>` is output. This
points to XpathRenderHandler::processValue*().
Replacing the tag with just lower case ascii immediatly solved the
problem. Seems a hiccup in an XML lib, because according to XML spec
underscore is perfectly fine. But this will make it a bit more robust