[BUGFIX] if node has same children multiple times it is not shown
Child nodes are retrieved and added to an array via a key <table>:<uid>
However this prevents the same content from being added to a container multiple times, so currently a FCE container that has the same CE 2 times will only show one element in the backend. As the view is out of sync with the flex field adding/moving/etc elements into and out of that container will create race conditions.