[BUGFIX] newCEwizard hook expects ExtendedNewContentElementController
the newCEwizard hook expects ExtendedNewContentElementController to retrieve the page.
This works if newCEwizard is called inside TVP, but not if its called in list module.

For the sake of not having a better fix, the whole filtering of FCEs is disabled in
list module to prevent an exception up until a better fix is available.